Add _save method to `attachment` post type actions
I've added a metabox to media attachments for a project I'm working on. At first, fields weren't saving. I discovered that the built-in "attachment" post type doesn't use the `save_post` action and instead uses `edit_attachment` and `add_attachment`. This patch fixes the saving issue and allows custom meta boxes to be attached to the `attachment` post type.